Using recursion for autoplay
I am trying to implement the speed and time delay for autoplay, however, the statement TimeUnit.SECONDS.sleep(...) was blocking the Swing Event Dispatch thread. This meant the game freezed and the player movements weren't being updated. Now I am trying to use recursion. This is the initial skeleton design, needs to be improved